Clean Project and Notes
Here's a 'clean' project along with the update notes.  This should
supersede any previous projects (to include the one I provided earlier
this evening.)  If you were to take the original 'JonColorWorks' project
and implement the changes on the notes, you should have a working
project with all changes made to date...hopefully.